+++ R/ddml_pliv.R
@@ -281,8 +281,8 @@ summary.ddml_pliv <- function(object, ...) {
 #' @rdname print.summary.ddml_plm
 #'
 #' @export
-print.summary.ddml_pliv <- function(x, ...) {
+print.summary.ddml_pliv <- function(x, digits = 3, ...) {
   cat("PLIV estimation results: \n \n")
   class(x) <- class(x)[-1]
-  print(x)
+  print(x, digits = digits)
 }#PRINT.SUMMARY.DDML_PLIV