+++ NEWS.md
@@ -1,5 +1,9 @@
 # ddml (development version)
 
+# ddml 0.2.2
+
+* Changes ``ddml:ols()`` default to ``const=TRUE``.
+* Adds probability forest compatibility to ``ddml::mdl_ranger()``.
 * Adds propensity score trimming option to ``ddml::ddml_ate()``, ``ddml::ddml_att()``, and ``ddml::ddml_late()``.
 * Fixes ATE and LATE scores.
 * Fixes output of ``ddml::print.summary.ddml_plm`` and ``ddml::print.summary.ddml_ate`` (#57).